+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+## cordova-plugin-sqlite-evplus-ext-common-free 0.6.0-dev
+
 ## cordova-plugin-sqlite-evplus-ext-common-free 0.5.0-pre01
 
 - updates from cordova-sqlite-evplus-ext-free-dependencies 0.2.0-rc01