+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## cordova-plugin-sqlite-evplus-ext-common-free 0.8.0-rc02
+
+- SQLite 3.36.0 update from cordova-sqlite-evplus-ext-free-dependencies@0.4.1
+
 ## cordova-plugin-sqlite-evplus-ext-common-free 0.8.0-rc01
 
 - use cordova-sqlite-evplus-ext-free-dependencies 0.4.0 with BLOBFROMBASE64 feature